feat: TUVA-134 Moved outlier data code to data file inside outlier project.
